Backend-Rezang-main/requirement.txt
absl-py==1.4.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
contourpy==1.0.7
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.10
django-allauth==0.54.0
django-cors-headers==4.2.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dlib==19.24.2
dnspython==2.4.1
executing==1.2.0
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
ipykernel==6.25.1
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
jupyter-client==8.3.0
jupyter-core==5.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mediapipe==0.10.1
multidict==6.0.4
mysqlclient==2.2.0
nest-asyncio==1.5.7
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
packaging==23.1
pandas==2.0.3
parso==0.8.3
phonenumbers==8.13.16
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pymongo==3.12.3
pyotp==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-snappy==0.6.1
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
secure-smtplib==0.1.1
six==1.16.0
sklearn==0.0.post7
sounddevice==0.4.6
sqlparse==0.2.4
stack-data==0.6.2
threadpoolctl==3.2.0
toml==0.10.2
tornado==6.3.2
traitlets==5.9.0
twilio==8.5.0
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.3.6
winkerberos==0.9.1
wrapt==1.12.1
yarl==1.9.2
zipp==3.15.0
